requirements.txt
anthropic>=0.40.0
slack-bolt>=1.18.0
slack-sdk>=3.26.0
python-dotenv>=1.0.0
